BackEnd/requirements.txt
aiobotocore==2.24.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
botocore==1.40.18
bounded-pool-executor==0.0.3
cachetools==6.2.0
certifi==2025.8.3
cftime==1.6.4.post1
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
Cython==3.1.4
dask==2025.9.1
earthaccess==0.15.1
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2025.9.0
google-auth==2.41.1
google-genai==1.41.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
locket==1.0.0
MarkupSafe==3.0.3
multidict==6.6.4
multimethod==2.0
narwhals==2.6.0
netCDF4==1.7.2
ngrok==1.5.1
numpy==1.26.4
packaging==25.0
pandas==2.3.3
partd==1.4.2
patsy==1.0.1
plotly==6.3.1
pmdarima==2.0.4
pqdm==0.2.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.10
pydantic_core==2.33.2
pyngrok==7.4.0
python-cmr==0.13.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
s3fs==2025.9.0
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
statsmodels==0.14.5
tenacity==9.1.2
threadpoolctl==3.6.0
tinynetrc==1.3.1
toolz==1.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xarray==2025.9.1
yarl==1.20.1
